Production statistics

+++ scripts/sort_hybrids.py
@@ -103,10 +103,10 @@ for hybrid in hybrid_list:
     print_text = "Hybrid: %s" % hybrid
     if int(production_data[21]) == 0:
         print_text += ", short circuit"
-    elif int(production_data[9]) != 0 and int(production_data[12]) == 0:
+    elif float(production_data[9]) != 0 and float(production_data[12]) == 0:
         print_text += ", failed S-curve"
     print print_text
-    print int(production_data[9])
-    print int(production_data[12])
+    print float(production_data[9])
+    print float(production_data[12])
     i  += 1
 print i
